test(network): simple transaction sync
